resolve deadlocks
When a lock is acquired and a other resources may acquire the same lock within a process.
Such as when performing attaching but a child process is added, then "Injector::attach" will be reentered, which leads to deadlocks.
The solution is to introduce threading when a child process arrives or session detaches.